+++ deode/suites/deode_suite_components.py
@@ -734,8 +734,9 @@ class LBCSubFamilyGenerator(EcflowSuiteFamily):
         self.limit = limit
         self.bdint = bdint
         self.member = member
-        self.do_slaf = do_slaf
-        if do_slaf:
+        self.do_glprep = self.config.get("suite_control.do_glprep", False)
+        self.do_slaf = do_slaf and not self.do_glprep
+        if self.do_slaf:
             # Must not exhaust the generator in the planning
             ltg1, ltg2 = tee(lbc_time_generator)
             self.lbc_time_generator = ltg1
@@ -747,11 +748,14 @@ class LBCSubFamilyGenerator(EcflowSuiteFamily):
     def __iter__(self):
         if self.do_slaf:
             bdshift = [as_timedelta("PT0H") for i in range(3)]
+        if self.config["suite_control.do_marsprep"]:
+            interpolation_task_name = "C903"
+        elif self.do_glprep:
+            interpolation_task_name = "GlBd"
+        else:
+            interpolation_task_name = "E927"
         for bd_index_time_dict in self.lbc_time_generator:
             bd_index_time_dict_sst = bd_index_time_dict.copy()
-            interpolation_task_name = (
-                "C903" if self.config["suite_control.do_marsprep"] else "E927"
-            )
             if (
                 self.config["suite_control.mode"] == "restart" and 0 in bd_index_time_dict
             ) or (