+++ deode/toolbox.py
@@ -8,8 +8,10 @@ import re
 import sys
 from typing import Any, Union
 
+import boto3
 import geohash
 import tomlkit
+from botocore.exceptions import ClientError
 from isodate import parse_duration
 from troika.connections.ssh import SSHConnection
 
@@ -266,26 +268,24 @@ class Platform:
             NotImplementedError: If provider not defined.
 
         """
-        # TODO handle platform differently archive etc
-        if provider_id == "symlink":
-            return LocalFileSystemSymlink(self.config, target, fetch=fetch)
+        providers = {
+            "symlink": LocalFileSystemSymlink,
+            "copy": LocalFileSystemCopy,
+            "move": LocalFileSystemMove,
+            "ecfs": ECFS,
+            "fdb": FDB,
+            "scp": SCP,
+            "s3": S3,
+        }
 
-        if provider_id == "copy":
-            return LocalFileSystemCopy(self.config, target, fetch=fetch)
+        try:
+            provider_cls = providers[provider_id]
+        except KeyError:
+            raise NotImplementedError(
+                f"Provider for {provider_id} not implemented"
+            ) from None
 
-        if provider_id == "move":
-            return LocalFileSystemMove(self.config, target, fetch=fetch)
-
-        if provider_id == "ecfs":
-            return ECFS(self.config, target, fetch=fetch)
-
-        if provider_id == "fdb":
-            return FDB(self.config, target, fetch=fetch)
-
-        if provider_id == "scp":
-            return SCP(self.config, target, fetch=fetch)
-
-        raise NotImplementedError(f"Provider for {provider_id} not implemented")
+        return provider_cls(self.config, target, fetch=fetch)
 
     def sub_value(self, pattern, key, value, micro="@", ci=True):
         """Substitute the value case-insensitively.
@@ -903,7 +903,9 @@ class FileManager:
             )
 
             logger.debug(
-                "Set output for target={} to destination={}", sub_target, sub_destination
+                "Set output for target={} to destination={}",
+                sub_target,
+                sub_destination,
             )
 
             logger.info("Checking archive provider_id {}", provider_id)
@@ -1396,6 +1398,74 @@ def compute_georef(domain_config):
     lon_center = domain_config["xloncen"]
 
     return geohash.encode(longitude=lon_center, latitude=lat_center, precision=6)
+
+
+class S3(ArchiveProvider):
+    """Transfer data with S3."""
+
+    def __init__(self, config, pattern, fetch=True):
+        """Construct S3 provider.
+
+        Args:
+            config (deode.ParsedConfig): Configuration
+            pattern (str): Filepattern
+            fetch (bool, optional): Fetch the data. Defaults to True.
+        """
+        ArchiveProvider.__init__(self, config, pattern, fetch=fetch)
+
+    def create_resource(self, resource):
+        """Create the resource.
+
+        Args:
+            resource (Resource): Resource.
+
+        Returns:
+            bool: True if success
+
+        Raises:
+            RuntimeError: If resource is not created
+
+        """
+        if "s3_endpoint_url" in self.config["system"]:
+            s3_endpoint_url = self.config["system.s3_endpoint_url"]
+        else:
+            raise RuntimeError(
+                "Error creating S3 provider, system.s3_endpoint_url missing in config"
+            )
+
+        try:
+            s3 = boto3.client("s3", endpoint_url=s3_endpoint_url)
+        except ClientError as e:
+            raise RuntimeError(
+                f"Error '{e}' creating S3 client for s3_endpoint_url={s3_endpoint_url}"
+            ) from None
+
+        if self.fetch:
+            logger.debug("s3 src={} to dst={}", self.identifier, resource.identifier)
+            parts = self.identifier.split("/")
+            s3_bucket_name = parts[0]
+            file_name = "/".join(parts[1:])
+            local_dir = os.path.dirname(resource.identifier)
+            if not os.path.isdir(local_dir):
+                os.makedirs(local_dir, mode=0o755, exist_ok=True)
+                logger.info("Created local directory {}", local_dir)
+            logger.debug(
+                "s3 download, s3_bucket_name={}, file_name={}, resource.identifier={}",
+                s3_bucket_name,
+                file_name,
+                resource.identifier,
+            )
+            s3.download_file(s3_bucket_name, file_name, resource.identifier)
+        else:
+            logger.debug("s3 src={} to dst={}", resource.identifier, self.identifier)
+            parts = self.identifier.split("/")
+            s3_bucket_name = parts[0]
+            file_name = "/".join(parts[1:])
+            # This is untested (Arcus is read only)
+            response = s3.upload_file(resource.identifier, s3_bucket_name, file_name)
+            logger.debug("s3 upload, response={}", response)
+
+        return True
 
 
 class Resource: